+++ taskchampion/src/taskstorage/mod.rs
@@ -80,10 +80,10 @@ pub trait TaskStorageTxn {
 
     /// Add a task to the working set and return its (one-based) index.  This index will be one greater
     /// than the highest used index.
-    fn add_to_working_set(&mut self, uuid: &Uuid) -> Fallible<u64>;
+    fn add_to_working_set(&mut self, uuid: &Uuid) -> Fallible<usize>;
 
     /// Remove a task from the working set.  Other tasks' indexes are not affected.
-    fn remove_from_working_set(&mut self, index: u64) -> Fallible<()>;
+    fn remove_from_working_set(&mut self, index: usize) -> Fallible<()>;
 
     /// Clear all tasks from the working set in preparation for a garbage-collection operation.
     fn clear_working_set(&mut self) -> Fallible<()>;
